backend/guardian/requirements.txt
pandas
requests
openai
torch
transformers
fastapi
uvicorn
plotly
python-dotenv